+++ templates/estragonia-godot/AvaloniaLoader.cs
@@ -1,4 +1,3 @@
-using Avalonia;
 using Godot;
 using JLeb.Estragonia;
 
@@ -8,12 +7,9 @@ namespace GodotGame;
 public partial class AvaloniaLoader : Node {
 
 	public override void _Ready() {
-		AppBuilder
-			.Configure<App>()
-			.UseGodot()
-			.SetupWithoutStarting();
+		if (!Engine.IsEditorHint())
+			GodotAvalonia.EnsureStarted<App>();
 
-		GodotAvalonia.EnsureAssetLoader(typeof(App).Assembly);
 		GetWindow()?.SetImeActive(true);
 	}